cleaup, minor reorg.
I noticed that 'ord' was identicial to the OrdSize record.  An
ordinal is literally just a record consiting of a carrier set
and a function from the carrier to ordinals! So the ordinal type
can play double duty as the type of ordinals, as well as the
canonical record associating a carrier type with its ordinal sizing
function.