Complete the definition of the comparison operation for BH stable forms.
It does not seem possible to write this function in the same nested
fixpoint style as previous ones. Instead, the computation is first characterized
via an inductive relation. We show the inductive relation is sufficient for
the corretness property. Then we use well-founded recursion to define
a function that computes according to the inductive relation.